requirements.txt
hyperopt
deap
tensorflow
transformers
gym
backtrader
xgboost
gensim
pytz
scikit-learn
pandas
scipy==1.10.1
textblob
matplotlib
torch
websocket-client